# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Game Week Voting API",
    "version": "0.1.0",
    "description": "Honor-system voter roster, one vote per award category, shared staff password for CSV upload and results."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque-signed-token",
        "description": "Session token from POST /sessions. Encodes voter_id and isStaff."
      }
    },
...
  },
  "paths": {
...
    "/sessions": {
      "post": {
        "summary": "Start a session by picking a roster name (optional shared staff password)",
        "tags": [
          "sessions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "displayName"
                ],
                "additionalProperties": false,
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "staffPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token",
                    "voterId",
                    "displayName",
                    "isStaff"
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "voterId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "isStaff": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games": {
      "get": {
        "summary": "List games that have not been withdrawn from the ballot",
        "tags": [
          "games"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "games"
                  ],
                  "properties": {
                    "games": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "gameId",
                          "title",
                          "submitterName",
                          "url"
                        ],
                        "properties": {
                          "gameId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": "string"
                          },
                          "submitterName": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ballot": {
      "get": {
        "summary": "Read the current voter's ballot draft or locked ballot",
        "tags": [
          "ballot"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ballotId",
                    "isLocked",
                    "lockedAt",
                    "votes"
                  ],
                  "properties": {
                    "ballotId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "uuid"
                    },
                    "isLocked": {
                      "type": "boolean"
                    },
                    "lockedAt": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "votes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "category",
                          "gameId"
                        ],
                        "properties": {
                          "category": {
                            "type": "string",
                            "enum": [
                              "technical_achievement",
                              "creative_or_fun_gameplay",
                              "visuals_or_graphics",
                              "best_overall"
                            ]
                          },
                          "gameId": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace the current voter's unlocked ballot draft",
        "tags": [
          "ballot"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "votes"
                ],
                "additionalProperties": false,
                "properties": {
                  "votes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "category",
                        "gameId"
                      ],
                      "properties": {
                        "category": {
                          "type": "string",
                          "enum": [
                            "technical_achievement",
                            "creative_or_fun_gameplay",
                            "visuals_or_graphics",
                            "best_overall"
                          ]
                        },
                        "gameId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ballotId",
                    "isLocked",
                    "lockedAt",
                    "votes"
                  ],
                  "properties": {
                    "ballotId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "uuid"
                    },
                    "isLocked": {
                      "type": "boolean"
                    },
                    "lockedAt": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "votes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "category",
                          "gameId"
                        ],
                        "properties": {
                          "category": {
                            "type": "string",
                            "enum": [
                              "technical_achievement",
                              "creative_or_fun_gameplay",
                              "visuals_or_graphics",
                              "best_overall"
                            ]
                          },
                          "gameId": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ballot/lock": {
      "post": {
        "summary": "Lock the current voter's ballot after all four categories are filled",
        "tags": [
          "ballot"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ballotId",
                    "isLocked",
                    "lockedAt",
                    "votes"
                  ],
                  "properties": {
                    "ballotId": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "uuid"
                    },
                    "isLocked": {
                      "type": "boolean"
                    },
                    "lockedAt": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "votes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "category",
                          "gameId"
                        ],
                        "properties": {
                          "category": {
                            "type": "string",
                            "enum": [
                              "technical_achievement",
                              "creative_or_fun_gameplay",
                              "visuals_or_graphics",
                              "best_overall"
                            ]
                          },
                          "gameId": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}